Time: 0 ms (100%), Space: 6.3 MB (6.97%) - LeetHub

+++ 2050-count-good-numbers/2050-count-good-numbers.cpp
@@ -24,6 +24,6 @@ int countGoodNumbers(long long n)
     long long first = (powe(5, even));
     long long second = (powe(4, odd));
 
-    return (int)((first * second) % mod);
+    return ((first * second) % mod);
 }
 };
